backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

motor
pymongo

python-jose[cryptography]
pyjwt
passlib[bcrypt]

pydantic
pydantic-settings

openai

fastembed
faiss-cpu

phonenumbers
email-validator

python-dotenv
requests
slowapi
